requirements.txt
fastapi
uvicorn
requests
beautifulsoup4
google-generativeai
python-jose
passlib[bcrypt]
slowapi
python-multipart
python-dotenv